Move test_utils to dbt_semantic_interfaces package
We're moving test_utils to dbt_sematnic_interfaces because they are needed
in both metricflow and DSI. Note though that we're moving them to the pacakge
library itself instead of the packages tests library. That is because we want
these test utils to be available to those implementing the the semantic
interfaces (whereas the tests themselves won't be bundled in distribution)
